requirements.txt
asgiref==3.8.1
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.1
django-bootstrap5==24.3
django-dbbackup==4.2.1
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
phonenumbers==8.13.47
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2024.2
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
